Refactor LibState to support partial initialization
The worker thread can be killed and respawned every time a new mission
is loaded. But due to limitations in egui/eframe, the gui thread cannot
be restarted. So I need to initialize the GUI thread only when the
library is loaded, but the worker thread once per DCS mission.

Also correctly join the worker thread so that it's actually destroyed
when a mission is unloaded.

Use an enum to handle the staged initialization. Thanks to mrkline for
the idea and most of the implementation.