Major refactor of logger messages
No longer send a separate message for every type of update. Instead,
one Update message with all of the info is sent to the logger's worker
thread.

Create an object for the logger so that I don't have to pass around a
ton of variables while splitting logger functionality into smaller
functions.